Use common config class
